+++ src/components/reusable/Swatch.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import styled, { css } from "styled-components";
-import { useHover } from "../hooks/useHover";
+import { useHover } from "../../hooks/useHover";
 import {
   ButtonDropdown,
   DropdownToggle,
